moved keystats api to root and changed import file path for those files removed local mock folders
