+++ src/__mock__/api.mock.ts
@@ -1,4 +1,5 @@
 import { Company } from '../types'
+import { mockKeyStats } from './apiData.mock'
 import { getCompanySymbols as companySymbols } from './companySymbols.mock'
 
 const mockFetch = (data: object) =>
@@ -12,3 +13,12 @@ export const getCompanySymbols = () => {
 }
 
 export default { getCompanySymbols }
+
+export const mockApi = {
+  getCompanySymbols: jest.fn().mockResolvedValue({}),
+  getHistory: jest.fn().mockResolvedValue({}),
+  getNews: jest.fn().mockResolvedValue({}),
+  getKeyStats: jest.fn().mockResolvedValue(mockKeyStats),
+  getOverview:jest.fn().mockResolvedValue({}),
+  getPeers: jest.fn().mockResolvedValue({})
+}
