+++ src/__mock__/apiData.mock.ts
@@ -51,3 +51,18 @@ export const historyInput1D = [
     minute: '13:10'
   }
 ]
+
+export const mockKeyStats = {
+  previousClose: 10,
+  dayHigh: 100,
+  dayLow: 10,
+  volume: 10000,
+  marketCap: 100000000,
+  peRatio: 10,
+  open: 10,
+  week52High: 100,
+  week52Low: 10,
+  avgTotalVolume: 10000,
+  earningsPerShare: 10,
+  dividendYield: 10
+}
