--- src/keystats/__tests__/__mock__/mockAPI.ts
@@ -1,10 +1,0 @@
-import { mockKeyStats } from './mockData'
-
-export default {
-  getCompanySymbols: jest.fn().mockResolvedValue({}),
-  getHistory: jest.fn().mockResolvedValue({}),
-  getNews: jest.fn().mockResolvedValue({}),
-  getKeyStats: jest.fn().mockResolvedValue(mockKeyStats),
-  getOverview:jest.fn().mockResolvedValue({}),
-  getPeers: jest.fn().mockResolvedValue({})
-}
