--- src/keystats/__tests__/__mock__/mockData.ts
@@ -1,14 +1,0 @@
-export const mockKeyStats = {
-  previousClose: 10,
-  dayHigh: 100,
-  dayLow: 10,
-  volume: 10000,
-  marketCap: 100000000,
-  peRatio: 10,
-  open: 10,
-  week52High: 100,
-  week52Low: 10,
-  avgTotalVolume: 10000,
-  earningsPerShare: 10,
-  dividendYield: 10
-}
