+++ src/keystats/__tests__/keystats.test.tsx
@@ -1,8 +1,8 @@
 import { shallow } from 'enzyme'
 import toJSON from 'enzyme-to-json'
 import React from 'react'
+import { mockKeyStats } from '../../__mock__/apiData.mock'
 import KeyStats from '../keystats'
-import { mockKeyStats } from './__mock__/mockData'
 
 describe('testing for keystats component', () => {
   it('should render key stats when keystats state updated', () => {
