+++ src/keystats/__tests__/keystatsAction.test.ts
@@ -1,8 +1,8 @@
 import { MockStore } from 'redux-mock-store'
+import { mockApi } from '../../__mock__/api.mock'
+import { mockKeyStats } from '../../__mock__/apiData.mock'
+import { generateMockStore } from '../../__mock__/mockStore.mock'
 import { getKeyStatsData, STATS_RECEIVED_ACTION, statsReceivedAction } from '../keystatsActions'
-import mockApi from './__mock__/mockAPI'
-import { mockKeyStats } from './__mock__/mockData'
-import { generateMockStore } from './__mock__/mockStore'
 
 const mockSymbol = 'aapl'
 
@@ -24,6 +24,7 @@ describe('test for keystats action', () => {
   describe('test for receive action', () => {
     it('should generate STATS_RECEIVED_ACTION action', () => {
       const action = statsReceivedAction(mockData)
+
       expect(action).toEqual(expectedAction)
     })
   })
@@ -31,9 +32,10 @@ describe('test for keystats action', () => {
   describe('test for async get thunk action', () => {
     it('should dispatch the STATS_RECEIVED_ACTION action when fetch is successful', async () => {
       const mockDispatch = store.dispatch
+      const mockStoreActions = store.getActions()
 
       await mockDispatch<any>(getKeyStatsData(mockSymbol))
-      const mockStoreActions = store.getActions()
+      
       expect(mockStoreActions[0]).toEqual(expectedAction)
     })
   })
