+++ src/keystats/__tests__/keystatsContainer.test.ts
@@ -1,6 +1,6 @@
 import { GlobalState } from 'types'
+import { mockKeyStats } from '../../__mock__/apiData.mock'
 import { mapStateToProps } from '../keystatsContainer'
-import { mockKeyStats } from './__mock__/mockData'
 
 describe('test for the mapStateToProps', () => {
   it('should return the keystats state', () => {
